refactor: use torchmetrics for BLEU and TER metrics
