+++ DashAI/back/metrics/translation/bleu.py
@@ -3,8 +3,7 @@
 from typing import TYPE_CHECKING
 
 from DashAI.back.core.utils import MultilingualString
-from DashAI.back.metrics.base_metric import prepare_to_metric
-from DashAI.back.metrics.translation_metric import TranslationMetric
+from DashAI.back.metrics.translation_metric import TranslationMetric, prepare_to_metric
 
 if TYPE_CHECKING:
     import numpy as np
@@ -21,6 +20,7 @@ class Bleu(TranslationMetric):
     References
     ----------
     - [1] https://en.wikipedia.org/wiki/BLEU
+    - [2] https://lightning.ai/docs/torchmetrics/stable/text/bleu_score.html
     """
 
     MAXIMIZE: bool = True
@@ -67,12 +67,17 @@ class Bleu(TranslationMetric):
         float
             The calculated BLEU score ranging between 0 and 1.
         """
-        import evaluate
+        from torchmetrics.text.bleu import BLEUScore
 
-        metric = evaluate.load("bleu")
+        bleu_metric = BLEUScore()
         source_sentences, target_sentences = prepare_to_metric(
-            source_sentences, target_sentences, "Bleu"
+            source_sentences, target_sentences
         )
-        return metric.compute(
-            references=source_sentences, predictions=target_sentences
-        )["bleu"]
+        return (
+            bleu_metric(
+                target_sentences,
+                [[reference] for reference in source_sentences],
+            )
+            .numpy()
+            .item()
+        )
