+++ DashAI/back/metrics/translation/ter.py
@@ -3,8 +3,7 @@
 from typing import TYPE_CHECKING
 
 from DashAI.back.core.utils import MultilingualString
-from DashAI.back.metrics.base_metric import prepare_to_metric
-from DashAI.back.metrics.translation_metric import TranslationMetric
+from DashAI.back.metrics.translation_metric import TranslationMetric, prepare_to_metric
 
 if TYPE_CHECKING:
     import numpy as np
@@ -20,7 +19,7 @@ class Ter(TranslationMetric):
 
     References
     ----------
-    - [1] https://huggingface.co/spaces/evaluate-metric/ter
+    - [1] https://lightning.ai/docs/torchmetrics/stable/text/translation_edit_rate.html
     """
 
     MAXIMIZE: bool = False
@@ -66,12 +65,17 @@ class Ter(TranslationMetric):
         float
             The calculated score.
         """
-        import evaluate
+        from torchmetrics.text.ter import TranslationEditRate
 
-        metric = evaluate.load("ter")
+        ter_metric = TranslationEditRate()
         source_sentences, target_sentences = prepare_to_metric(
-            source_sentences, target_sentences, "Ter"
+            source_sentences, target_sentences
         )
-        return metric.compute(
-            references=source_sentences, predictions=target_sentences
-        )["score"]
+        return (
+            ter_metric(
+                target_sentences,
+                source_sentences,
+            )
+            .numpy()
+            .item()
+        )
